flask-project/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
flask_sqlalchemy==2.5.1
Flask-Login==0.6.2
flask_wtf==1.0.0
wtforms==3.0.1
pymongo==4.15.4
bcrypt==5.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
dnspython==2.7.0
flask_mail==0.9.1
